+++ src/archetype.rs
@@ -824,10 +824,10 @@ mod tests {
 
         world.insert(e, C("hello".into()));
 
-        assert_eq!(world.get_component::<C>(e).unwrap().0, "hello");
+        assert_eq!(world.get::<C>(e).unwrap().0, "hello");
 
         world.insert(e, C("goodbye".into()));
 
-        assert_eq!(world.get_component::<C>(e).unwrap().0, "goodbye");
+        assert_eq!(world.get::<C>(e).unwrap().0, "goodbye");
     }
 }